+++ auditlog/diff.py
@@ -79,7 +79,10 @@ def get_field_value(obj, field):
                 value = django_timezone.make_naive(value, timezone=timezone.utc)
         elif isinstance(field, JSONField):
             value = field.to_python(getattr(obj, field.name, None))
-            value = json.dumps(value, sort_keys=True, cls=field.encoder)
+            try:
+                value = json.dumps(value, sort_keys=True, cls=field.encoder)
+            except TypeError:
+                pass
         elif (field.one_to_one or field.many_to_one) and hasattr(field, "rel_class"):
             value = smart_str(
                 getattr(obj, field.get_attname(), None), strings_only=True