+++ auditlog/context.py
@@ -13,12 +13,13 @@ auditlog_disabled = ContextVar("auditlog_disabled", default=False)
 
 
 @contextlib.contextmanager
-def set_actor(actor, remote_addr=None):
+def set_actor(actor, remote_addr=None, remote_port=None):
     """Connect a signal receiver with current user attached."""
     # Initialize thread local storage
     context_data = {
         "signal_duid": ("set_actor", time.time()),
         "remote_addr": remote_addr,
+        "remote_port": remote_port,
     }
     auditlog_value.set(context_data)
 
@@ -63,6 +64,7 @@ def _set_actor(user, sender, instance, signal_duid, **kwargs):
             instance.actor = user
 
         instance.remote_addr = auditlog["remote_addr"]
+        instance.remote_port = auditlog["remote_port"]
 
 
 @contextlib.contextmanager