+++ auditlog/conf.py
@@ -46,10 +46,7 @@ settings.AUDITLOG_DISABLE_REMOTE_ADDR = getattr(
     settings, "AUDITLOG_DISABLE_REMOTE_ADDR", False
 )
 
-# Enable changes_display_dict truncator
-settings.AUDITLOG_TRUNCATE_CHANGES_DISPLAY = getattr(
-    settings, "AUDITLOG_TRUNCATE_CHANGES_DISPLAY", True
-)
-
 # Number of characters at which changes_display_dict property should be shown
-settings.AUDITLOG_TRUNCATE_LIMIT = getattr(settings, "AUDITLOG_TRUNCATE_LIMIT", 140)
+settings.AUDITLOG_CHANGE_DISPLAY_TRUNCATE_LENGTH = getattr(
+    settings, "AUDITLOG_CHANGE_DISPLAY_TRUNCATE_LENGTH", 140
+)