+++ auditlog/models.py
@@ -24,7 +24,6 @@ from django.utils.encoding import smart_str
 from django.utils.translation import gettext_lazy as _
 
 from auditlog.diff import mask_str
-from auditlog.text import truncatechars
 
 DEFAULT_OBJECT_REPR = "<error forming object repr>"
 
@@ -507,11 +506,9 @@ class LogEntry(models.Model):
                     elif field_type in ["ForeignKey", "OneToOneField"]:
                         value = self._get_changes_display_for_fk_field(field, value)
 
-                    if (
-                        settings.AUDITLOG_TRUNCATE_CHANGES_DISPLAY
-                        and len(value) > settings.AUDITLOG_TRUNCATE_LIMIT
-                    ):
-                        value = truncatechars(value, settings.AUDITLOG_TRUNCATE_LIMIT)
+                    truncate_at = settings.AUDITLOG_CHANGE_DISPLAY_TRUNCATE_LENGTH
+                    if 0 <= truncate_at < len(value):
+                        value = value[:truncate_at] + ("..." if truncate_at > 0 else "")
 
                     values_display.append(value)